LayoutTests/imported/w3c/web-platform-tests/wasm/jsapi/memory/to-fixed-length-buffer.any-expected.txt


FAIL API surface assert_equals: expected "object" but got "undefined"
FAIL toFixedLengthBuffer caching behavior memory.toFixedLengthBuffer is not a function. (In 'memory.toFixedLengthBuffer()', 'memory.toFixedLengthBuffer' is undefined)
